Agentic-AI-Course/agenticappbackend/requirements.txt
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==46.0.5
distro==1.9.0
Django==6.0.2
django-cors-headers==4.9.0
django-ninja==1.5.3
django-ninja-extra==0.31.0
django-ninja-jwt==5.4.4
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
injector==0.24.0
jiter==0.13.0
openai==2.21.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
